+++ src/lang.c
@@ -1,0 +1,6 @@
+#define R_NO_REMAP
+#include <Rinternals.h>
+
+SEXP rlang_new_language(SEXP head, SEXP tail) {
+  return Rf_lcons(head, tail);
+}